Fix bitcoind startup error handling
Like Bisq instance startup, bitcoind startup should save any error msg
from bash, print the error msg, and System.exit(1).
